SSH retry loop
Customers have asked for a way to retry the SSH portion of creating a server. Basically we don't know what they're installing/waiting for on the server so SSH may or may not wait long enough and otherwise err out. This patch introduces `ssh_retry_every` option as a timer for a retry loop.